requirements.txt
asgiref==3.8.1
Django==5.0.6
django-extensions==3.2.3
django-scripts==2.0
djangorestframework==3.15.1
geographiclib==2.0
geopy==2.4.1
install==1.3.5
sqlparse==0.5.0
stua==0.3
